vigia-anemia-api/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# FHIR Support
fhir.resources==7.1.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS Integration
boto3==1.29.7
botocore==1.32.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Caching
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Validation
email-validator==2.1.0
validate-docbr==1.10.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pylint==3.0.3

# Development
ipython==8.18.1
ipdb==0.13.13

